+++ hc/front/urls.py
@@ -77,6 +77,7 @@ project_urls = [
     path("add_zulip/", views.add_zulip, name="hc-add-zulip"),
     path("add_spike/", views.add_spike, name="hc-add-spike"),
     path("add_linenotify/", views.add_linenotify, name="hc-add-linenotify"),
+    path("add_signal/", views.add_signal, name="hc-add-signal"),
     path("badges/", views.badges, name="hc-badges"),
     path("checks/", views.my_checks, name="hc-checks"),
     path("checks/add/", views.add_check, name="hc-add-check"),